+++ lib/domain/auth/auth_repository.dart
@@ -1,0 +1,9 @@
+import 'auth_user.dart';
+
+abstract interface class AuthRepository {
+  Stream<AuthUser?> get authStateChanges;
+
+  Future<AuthUser> signInWithGoogle();
+
+  Future<void> signOut();
+}